Refactor: Consolidate signature model logic
Merge real_signature_model into signature_model and update API endpoints to use the unified model.
